+++ src/extension.ts
@@ -1,12 +1,14 @@
 import * as vscode from 'vscode';
-import { newViewModel } from './commands/index'
+import { newProject, newViewModel } from './commands/index'
 
 export function activate(context: vscode.ExtensionContext) {
 	console.log('Extension "mvvm-flutter" is now active!');
 
-	const disposable = vscode.commands.registerCommand('mvvm-flutter.new-view-model', newViewModel);
-
-	context.subscriptions.push(disposable);
+	context.subscriptions.push(
+		vscode.commands.registerCommand('mvvm-flutter.new-view-model', newViewModel),
+		vscode.commands.registerCommand('mvvm-flutter.new-project',
+			() => newProject(context.extensionPath))
+	);
 }
 
 export function deactivate() { }