+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to the "mvvm-flutter" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [v1.1.0]
 
-- Initial release
+### Added
+
+- Project generation
+
+### Fixed
+
+- New ViewModel typos
+
+## [v1.0.0]
+
+### Added
+
+- Initial release, ViewModel files generation